+++ xr_3da/Occluder.cpp
@@ -46,7 +46,7 @@ void COccluderSystem::Load(CStream *fs)
 
 		// register
 		SOccluderDef	def;
-		for (int j=0; j<L.vertices.size(); j++)
+		for (size_t j=0; j<L.vertices.size(); j++)
 			def.verts.push_back	(L.vertices[j]);
 		def.sphere.compute		(def.verts.begin(),def.verts.size());
 		occluders.push_back	(def);
